그림자 depthonly pass 적

+++ D3D_Practice_9thWeek/Animation.cpp
@@ -1,9 +1,9 @@
 #include "Animation.h"
 
-void BoneAnimation::Evaluate(float time, Vector3& position, Quaternion& rotation, Vector3& scale)
+void BoneAnimation::Evaluate(double time, Vector3& position, Quaternion& rotation, Vector3& scale)
 {
 	// 키프레임 시간이 아니라, 델타타임 기준으로 트랜스폼 값 보간
-	float currentTime = fmod(time, AnimationKeys.back().Time);
+	double currentTime = fmod(time, AnimationKeys.back().Time);
 	int frameIndex = 0;
 	for (int i = 0; i < AnimationKeys.size()-1; i++)
 	{
@@ -15,10 +15,10 @@ void BoneAnimation::Evaluate(float time, Vector3& position, Quaternion& rotation
 	}
 	int nextIndex = frameIndex + 1;
 
-	float deltaFrame = AnimationKeys[nextIndex].Time - AnimationKeys[frameIndex].Time;
-	float rate = (currentTime- AnimationKeys[frameIndex].Time) / deltaFrame;
+	double deltaFrame = AnimationKeys[nextIndex].Time - AnimationKeys[frameIndex].Time;
+	double rate = (currentTime- AnimationKeys[frameIndex].Time) / deltaFrame;
 	
-	position = Vector3::Lerp(AnimationKeys[frameIndex].Position, AnimationKeys[nextIndex].Position, rate);
-	rotation = Quaternion::Slerp(AnimationKeys[frameIndex].Rotation, AnimationKeys[nextIndex].Rotation, rate);
-	scale = Vector3::Lerp(AnimationKeys[frameIndex].Scale, AnimationKeys[nextIndex].Scale, rate);
+	position = Vector3::Lerp(AnimationKeys[frameIndex].Position, AnimationKeys[nextIndex].Position, (float)rate);
+	rotation = Quaternion::Slerp(AnimationKeys[frameIndex].Rotation, AnimationKeys[nextIndex].Rotation, (float)rate);
+	scale = Vector3::Lerp(AnimationKeys[frameIndex].Scale, AnimationKeys[nextIndex].Scale, (float)rate);
 }
