+++ D3D_Practice_9thWeek/Animation.h
@@ -5,7 +5,7 @@ using namespace DirectX::SimpleMath;
 
 struct KeyFrame // 프레임 하나하나
 {
-	float Time;
+	double Time;
 	Vector3 Position;
 	Quaternion  Rotation;
 	Vector3 Scale;
@@ -16,7 +16,7 @@ struct BoneAnimation // 프레임들을 모아둔 본 한개짜리의 애니메이션
 	std::vector<KeyFrame> AnimationKeys;
 	std::string m_Name;
 	//보간을 해서 프레임 사이의 변환값 만들기
-	void Evaluate(float time, Vector3& position, Quaternion& rotation, Vector3& scale);
+	void Evaluate(double time, Vector3& position, Quaternion& rotation, Vector3& scale);
 };
 
 struct Animation // 모든 본의 애니메이션 정보를 모아둔 애니메이션 그 자체
